+++ Konsole/IConsole.cs
@@ -15,6 +15,10 @@ namespace Konsole
         void PrintAt(int x, int y, string format, params object[] args);
         void PrintAt(int x, int y, string text);
         void PrintAt(int x, int y, char c);
+        void ScrollUp();
         void Clear();
+
+        void MoveBufferArea(int sourceLeft, int sourceTop, int sourceWidth, int sourceHeight, int targetLeft,
+            int targetTop, char sourceChar, ConsoleColor sourceForeColor, ConsoleColor sourceBackColor);
     }
 }